+++ server/src/backoffice/telegram_logs/sender.py
@@ -2,6 +2,8 @@ from abc import ABC, abstractmethod
 
 import structlog
 from aiogram import Bot
+from aiogram.client.default import DefaultBotProperties
+from aiogram.enums import ParseMode
 
 from src.config import settings
 from src.enums import TelegramLogSender as TelegramLogSenderType
@@ -25,7 +27,10 @@ class LoggingTelegramLogSender(BaseTelegramLogSender):
 
 class TelegramLogSender(BaseTelegramLogSender):
     def __init__(self) -> None:
-        self.bot = Bot(token=settings.BOT_TOKEN)
+        self.bot = Bot(
+            token=settings.BOT_TOKEN,
+            default=DefaultBotProperties(parse_mode=ParseMode.HTML),
+        )
 
     async def send(self, text: str, with_notification: bool) -> None:
         await self.bot.send_message(